requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.2
cryptography==46.0.7
decorator==5.2.1
discord==2.3.2
discord.py==2.5.2
distro==1.9.0
docstring_parser==0.18.0
executing==2.2.0
fastapi==0.115.12
filelock==3.28.0
Flask==3.1.0
frozenlist==1.5.0
fsspec==2026.3.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-cloud-aiplatform==1.148.1
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-cloud-resource-manager==1.17.0
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.73.1
google-generativeai==0.8.6
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
gradio_client==2.4.1
grpc-google-iam-v1==0.14.4
grpcio==1.80.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.10
iniconfig==2.1.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.3.2
narwhals==1.33.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
plotly==6.0.1
pluggy==1.5.0
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.27.2
protobuf==5.29.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.11.2
pydantic_core==2.33.1
pygame==2.6.1
Pygments==2.19.1
pyparsing==3.3.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.3
rich==15.0.0
scikit-learn==1.6.1
scipy==1.15.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-inspection==0.4.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==16.0
Werkzeug==3.1.3
yarl==1.19.0
